kernel: use _fast8_t types

+++ Kernel/filesys.c
@@ -21,8 +21,8 @@
 
 uint8_t lastname[31];
 
-static uint8_t n_open_fault;
-static uint8_t n_fault_type;
+static uint_fast8_t n_open_fault;
+static uint_fast8_t n_fault_type;
 static uint8_t *name, *nameend;
 
 static uint8_t getcf(void)
@@ -199,7 +199,7 @@ inoptr srch_dir(inoptr wd, uint8_t *compname)
 
 inoptr srch_mt(inoptr ino)
 {
-    uint8_t j;
+    uint_fast8_t j;
     struct mount *m = &fs_tab[0];
 
     for(j=0; j < NMOUNTS; ++j){
@@ -417,7 +417,7 @@ bool ch_link(inoptr wd, uint8_t *oldname, uint8_t *newname, inoptr nindex)
  */
 bool namecomp(uint8_t *n1, uint8_t *n2) // return true if n1 == n2
 {
-    uint8_t n; // do we have enough variables called n?
+    uint_fast8_t n; // do we have enough variables called n?
 
     n = FILENAME_LEN;
     while(*n1 && *n1 != '/')
@@ -449,7 +449,7 @@ bool namecomp(uint8_t *n1, uint8_t *n2) // return true if n1 == n2
 inoptr newfile(inoptr pino, uint8_t *name)
 {
     regptr inoptr nindex;
-    uint8_t j;
+    uint_fast8_t j;
 
     /* No parent? */
     if (!pino) {
@@ -735,9 +735,9 @@ void blk_free(uint16_t devno, blkno_t blk)
  * entries in the open file table.
  */
 
-int8_t oft_alloc(void)
+int_fast8_t oft_alloc(void)
 {
-    uint8_t j;
+    uint_fast8_t j;
 
     for(j=0; j < OFTSIZE ; ++j) {
         if(of_tab[j].o_refs == 0) {
@@ -759,7 +759,7 @@ int8_t oft_alloc(void)
 void deflock(regptr struct oft *ofptr)
 {
     inoptr i = ofptr->o_inode;
-    uint8_t c = i->c_flags & CFLOCK;
+    uint_fast8_t c = i->c_flags & CFLOCK;
 
     if (ofptr->o_access & O_FLOCK) {
         if (c == CFLEX)
@@ -776,7 +776,7 @@ void deflock(regptr struct oft *ofptr)
  *	from a user file table then drop any file locks, dereference the inode
  *	and mark empty
  */
-void oft_deref(int8_t of)
+void oft_deref(uint_fast8_t of)
 {
     struct oft *ofptr;
 
@@ -791,9 +791,9 @@ void oft_deref(int8_t of)
 
 /* Uf_alloc finds an unused slot in the user file table.*/
 
-int8_t uf_alloc_n(int base)
+int_fast8_t uf_alloc_n(uint_fast8_t base)
 {
-    uint8_t j;
+    uint_fast8_t j;
 
     for(j=base; j < UFTSIZE ; ++j) {
         if(udata.u_files[j] == NO_FILE) {
@@ -805,7 +805,7 @@ int8_t uf_alloc_n(int base)
 }
 
 
-int8_t uf_alloc(void)
+int_fast8_t uf_alloc(void)
 {
     return uf_alloc_n(0);
 }
@@ -818,7 +818,7 @@ int8_t uf_alloc(void)
 
 void i_deref(regptr inoptr ino)
 {
-    uint8_t mode = getmode(ino);
+    uint_fast8_t mode = getmode(ino);
 
     magic(ino);
 
@@ -899,7 +899,7 @@ uint16_t devnum(inoptr ino)
 int f_trunc(regptr inoptr ino)
 {
     uint16_t dev;
-    int8_t j;
+    int_fast8_t j;
 
     if (ino->c_flags & CRDONLY) {
         udata.u_error = EROFS;
@@ -933,7 +933,7 @@ int f_trunc(regptr inoptr ino)
    This is annoying and it would be nice one day to find a clean solution */
 
 #ifdef CONFIG_BLKBUF_EXTERNAL
-void freeblk(uint16_t dev, blkno_t blk, uint8_t level)
+void freeblk(uint16_t dev, blkno_t blk, uint_fast8_t level)
 {
     struct blkbuf *buf;
     regptr blkno_t *bn;
@@ -959,7 +959,7 @@ void freeblk(uint16_t dev, blkno_t blk, uint8_t level)
 
 #else
 
-void freeblk(uint16_t dev, blkno_t blk, uint8_t level)
+void freeblk(uint16_t dev, blkno_t blk, uint_fast8_t level)
 {
     struct blkbuf *buf;
     regptr blkno_t *bn;
@@ -989,7 +989,7 @@ void freeblk(uint16_t dev, blkno_t blk, uint8_t level)
  * the physical block number on a device given the inode and the
  * logical block number in a file.  The block is zeroed if created.
  */
-blkno_t bmap(inoptr ip, blkno_t bn, int rwflg)
+blkno_t bmap(inoptr ip, blkno_t bn, unsigned int rwflg)
 {
     int i;
     bufptr bp;
@@ -1114,11 +1114,11 @@ static const uint16_t blkmask[] = {
     0x1F	/* 16384 */
 };
 
-blkno_t bmap(inoptr ip, blkno_t blk, int rwflg)
+blkno_t bmap(inoptr ip, blkno_t blk, unsigned int rwflg)
 {
     /* Linear bits */
-    uint8_t shift = fs_tab[ip->c_super].m_fs.s_shift;
-    uint8_t blklo = ((uint8_t)blk) & blkmask[shift];
+    uint_fast8_t shift = fs_tab[ip->c_super].m_fs.s_shift;
+    uint_fast8_t blklo = ((uint8_t)blk) & blkmask[shift];
     /* Non linear index bits */
     blk >>= shift;
     blk = do_bmap(ip, blk, rwflg);
@@ -1153,9 +1153,9 @@ void validblk(uint16_t dev, blkno_t num)
 /* This returns the inode pointer associated with a user's file
  * descriptor, checking for valid data structures.
  */
-inoptr getinode(uint8_t uindex)
+inoptr getinode(uint_fast8_t uindex)
 {
-    uint8_t oftindex;
+    uint_fast8_t oftindex;
     inoptr inoindex;
 
     if(uindex >= UFTSIZE || udata.u_files[uindex] == NO_FILE) {
@@ -1218,7 +1218,7 @@ uint8_t getperm(inoptr ino)
 
 
 /* This sets the times of the given inode, according to the flags. */
-void setftime(inoptr ino, uint8_t flag)
+void setftime(inoptr ino, uint_fast8_t flag)
 {
     if (ino->c_flags & CRDONLY)
         return;
@@ -1245,7 +1245,7 @@ uint8_t getmode(inoptr ino)
 static struct mount *newfstab(void)
 {
     struct mount *m = fs_tab;
-    int i;
+    uint_fast8_t i;
     for (i = 0; i < NMOUNTS; i++) {
         if (m->m_dev == NO_DEVICE)
             return m;
